Split release checker
